# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SKF Racing Hub API",
    "description": "REST API of the SKF Racing Hub site.\n\n**Authentication.** A `session_id` cookie, set by the Discord login. Operations\nthat need a session list it under *security*; the roles they accept are in\n`x-required-roles`.\n\n**Errors.** Every error is an `application/problem+json` document (`Problem`).\n`code` is a stable machine-readable identifier, `detail` a message that can be\nshown to the user. Validation failures (422) also carry `errors`, one entry per\nrejected field.\n\n**Paging.** Collections that take `limit` and `offset` return a plain array;\nthe total is in `X-Total-Count` and neighbouring pages in an RFC 8288 `Link`\nheader.\n\n**Tracing.** Every response carries `X-Request-ID`; quote it when reporting a\nproblem.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/penalty-rules": {
      "get": {
        "tags": [
          "Penalty rules"
        ],
        "summary": "List penalty rules",
        "operationId": "listPenaltyRules",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PenaltyRuleOut"
                  },
                  "type": "array"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Penalty rules"
        ],
        "summary": "Create penalty rule",
        "operationId": "createPenaltyRule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PenaltyRuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PenaltyRuleOut"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Path of the created resource.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      }
    },
    "/api/v1/penalty-rules/{ruleId}": {
      "get": {
        "tags": [
          "Penalty rules"
        ],
        "summary": "Get penalty rule",
        "operationId": "getPenaltyRule",
        "parameters": [
          {
            "name": "ruleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PenaltyRuleOut"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        }
      },
      "patch": {
        "tags": [
          "Penalty rules"
        ],
        "summary": "Update penalty rule",
        "operationId": "updatePenaltyRule",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "ruleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PenaltyRuleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PenaltyRuleOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      },
      "delete": {
        "tags": [
          "Penalty rules"
        ],
        "summary": "Delete penalty rule",
        "operationId": "deletePenaltyRule",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "ruleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "type": "object"
      },
...
      "PenaltyRuleCreate": {
        "properties": {
          "threshold": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "label": {
            "type": "string",
            "maxLength": 500,
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "threshold"
        ]
      },
      "PenaltyRuleOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "threshold": {
            "type": "integer"
          },
          "label": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer"
          }
        },
        "type": "object",
        "required": [
          "id",
          "threshold",
          "label",
          "sortOrder"
        ]
      },
      "PenaltyRuleUpdate": {
        "properties": {
          "threshold": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "label": {
            "type": "string",
            "maxLength": 500
          }
        },
        "type": "object"
      },
      "Problem": {
        "description": "An RFC 7807 problem document. Every error response has this shape.",
        "example": {
          "code": "not_found",
          "detail": "Driver not found.",
          "status": 404,
          "title": "Not Found",
          "type": "about:blank"
        },
        "properties": {
          "type": {
            "default": "about:blank",
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "type": "object"
      },
...
      "ValidationProblem": {
        "description": "A 422: the request did not pass validation.",
        "example": {
          "code": "validation_error",
          "detail": "year: Field required",
          "errors": [
            {
              "field": "year",
              "message": "Field required"
            }
          ],
          "status": 422,
          "title": "Unprocessable Content",
          "type": "about:blank"
        },
        "properties": {
          "type": {
            "default": "about:blank",
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "errors"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Session id, set as a cookie by the Discord login.",
        "in": "cookie",
        "name": "session_id"
      },
      "ingestToken": {
        "type": "http",
        "description": "Shared token of the external incident parsers.",
        "scheme": "bearer"
      }
    },
    "responses": {
...
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "UnprocessableContent": {
        "description": "Unprocessable Content",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationProblem"
            }
          }
        }
      }
    }
  },
...
}